+++ core/matching/order_status_test.go
@@ -806,12 +806,13 @@ type marketPositionFake struct {
 	party string
 }
 
-func (m marketPositionFake) Party() string             { return m.party }
-func (m marketPositionFake) Size() int64               { return 0 }
-func (m marketPositionFake) Buy() int64                { return 0 }
-func (m marketPositionFake) Sell() int64               { return 0 }
-func (m marketPositionFake) Price() *num.Uint          { return num.UintZero() }
-func (m marketPositionFake) BuySumProduct() *num.Uint  { return num.UintZero() }
-func (m marketPositionFake) SellSumProduct() *num.Uint { return num.UintZero() }
-func (m marketPositionFake) VWBuy() *num.Uint          { return num.UintZero() }
-func (m marketPositionFake) VWSell() *num.Uint         { return num.UintZero() }
+func (m marketPositionFake) AverageEntryPrice() *num.Uint { return num.UintZero() }
+func (m marketPositionFake) Party() string                { return m.party }
+func (m marketPositionFake) Size() int64                  { return 0 }
+func (m marketPositionFake) Buy() int64                   { return 0 }
+func (m marketPositionFake) Sell() int64                  { return 0 }
+func (m marketPositionFake) Price() *num.Uint             { return num.UintZero() }
+func (m marketPositionFake) BuySumProduct() *num.Uint     { return num.UintZero() }
+func (m marketPositionFake) SellSumProduct() *num.Uint    { return num.UintZero() }
+func (m marketPositionFake) VWBuy() *num.Uint             { return num.UintZero() }
+func (m marketPositionFake) VWSell() *num.Uint            { return num.UintZero() }